pyproject.toml
[tool.poetry]
name = "prison-book-project"
version = "0.1.0"
description = ""
authors = ["ges-na <gesinaface@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django-environ = "^0.8.1"
Django = "^4.0.4"
# psycopg2 = "^2.9.3"
django-extensions = "^3.1.5"
django-import-export = "^2.8.0"
psycopg2-binary = "^2.9.3"
model-bakery = "^1.5.0"
ipdb = "^0.13.9"
django-cors-headers = "^3.13.0"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"
django-ajax-selects = "^2.2.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.400"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["app/*", "auth/*", "prisonbookproject/*", "tests/"]
exclude = [
    "README.md",
    ".venv",
]
useLibraryCodeForTypes = true
reportIncompatibleVariableOverride = false
reportIncompatibleMethodOverride = false
venvPath = "."
venv = ".venv"
stubPath = ""

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
extend-safe-fixes = ["I"]
extend-select = ["I"]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = true
combine-as-imports = true
known-third-party = ["django","django.contrib","django.utils"]
